feat: use own authentication instead of authjs
This implements the authentication method provided by our rust backend
since authjs is a pain in the ass to work with for our use case.
Currently our method is to check for auth in every page load which is a
bit ugly IMO. I don't fully understand how `hooks` work in svelte and if
it supports SSG so we'll stick with the current method.