feat: add own authentication feature
Instead of relying on authjs to issue JWTs we now do it ourselves. We
now return a JWT with 24 hours expiry on successful authentication. We
also provide a new `get_session` API which handles verification of
tokens stored in localStorage on client side, on successful
verification we respond with the claims of the token.