feat: add navbar functionalities and required routes (#5)
* feat: make navigation bar functional

This makes the navigation bar functional by updating the dropdown items
to reflect our use case. We make the items point to their respective
pages. We also added profile and logout dropdown items on the navbar end
button which are both functional.

* feat: add logout function

This handles the removal of `auth_token` in the localStorage and also
introducing a synthetic 1 second delay. We also increased the delay from
500msec to 1 sec in our `goto` workaround.

* feat: make loading message more concise

* feat: update page header

* feat: add required pages

This adds all the required routes in our project. Currently all of the
pages only contains templates.