feat: initial feedback page implementation
This creates the feedback page template. Currently this only implements
the auth checking along with the prevention of navigating back to
dashboard route. By design the only way in to this page is through the
dashboard, since we are using Tauri the user can't manipulate the
link. We might need to handle navigating back to the results page.