feat: make the rating data usable
This makes the rating data accessible by binding groups to each of the
radio button along with the proper numerical values. Also added the
feature that when the client allowed face recording the backend will be
notified by invoking the `start_face_recording` function. The form data
now also get's sent to the backend when the user clicks the submit
button.