feat: implement face recording
This implements the face recording feature by calling the recorder
module. This module contains 3 functions which just wraps the calls to
the python backend. We call the `recorder::start` in the
start_face_recording command and handle the result using pattern
matching. We call `recorder::stop` when in `submit_feedback` which is
called when the user submits the form. Last is we added a new command
`clear_recording` which calls `recorder::clear`, this get's invoked when
the user returns to the consent screen.