feat: add python backend for recording
Since threading in rust is quite hard for me, I decided to just use
python to implement the recording feature. This is done by exposing an
endpoint which calls to internal methods that handles the spawning of
threads to run ffmpeg in record mode and also providing a way to
terminate the thread by calling a different endpoint.