feat: thank you modal and guarantee emotion data completeness (#19)
* feat: add new component ThankYouModal

This replaces the `toast.success` we use to display our thank you
message.

closes #15

* feat: use modal and prevent incomplete emotion data

This uses the new component `ThankYouModal` to display the thank you
message. Also implements a check to ensure all emotion data is resolved
before allowing submission.

closes #18

* feat: stop analysis instead of clear recording

We're not recording anymore so we will just stop the analysis when clear
recording method is called. Too lazy to name them properly, they work
and that's it.