feat: update feedback tables schema, allow filter
This updates both the trad_feedback_data and hybrid_feedback_data tables
schema. This is done to allow us to easily filter based on start and end
dates. This also introduces the new function get_filtered_feedbacks
which allows us to query matching rows based on a start and end date.