feat: implement archiving process
This commit introduces two new tables prefixed with archived_x where x
corresponds to the original table where the data is from. We implemented
the archiving by querying the original table for rows where the
created_at row subtracted to the current unix timestamp is greater that
5 years in seconds. We will then move the rows to the archived_x table
then delete rows in the original table.