feat: use mysql db for feedback data
This moves all the feedback related data to the new central mysql db
instead of the local sqlite db. The sqlite db will be kept and will be
utilized to store user specific data such as services and credentials.

NOTE: the connection string to mysql db is set with MYSQL_DB_URL
envvar.